+++ pom.xml
@@ -110,7 +110,6 @@
 	    <dependency>
 	        <groupId>com.h2database</groupId>
 	        <artifactId>h2</artifactId>
-	        <scope>test</scope>
 	    </dependency>		
 		<dependency>
 			<groupId>org.apache.tomcat</groupId>
