+++ src/main/java/business/security/CsrfHeaderFilter.java
@@ -22,17 +22,16 @@ public class CsrfHeaderFilter extends OncePerRequestFilter {
     protected void doFilterInternal(HttpServletRequest request,
             HttpServletResponse response, FilterChain filterChain)
             throws ServletException, IOException {
-      CsrfToken csrf = (CsrfToken) request.getAttribute(CsrfToken.class
-          .getName());
-      if (csrf != null) {
-        Cookie cookie = WebUtils.getCookie(request, "XSRF-TOKEN");
-        String token = csrf.getToken();
-        if (cookie==null || token!=null && !token.equals(cookie.getValue())) {
-          cookie = new Cookie("XSRF-TOKEN", token);
-          cookie.setPath("/");
-          response.addCookie(cookie);
+        CsrfToken csrf = (CsrfToken)request.getAttribute(CsrfToken.class.getName());
+        if (csrf != null) {
+            Cookie cookie = WebUtils.getCookie(request, "XSRF-TOKEN");
+            String token = csrf.getToken();
+            if (cookie == null || token != null && !token.equals(cookie.getValue())) {
+                cookie = new Cookie("XSRF-TOKEN", token);
+                cookie.setPath("/");
+                response.addCookie(cookie);
+            }
         }
-      }
-      filterChain.doFilter(request, response);
+        filterChain.doFilter(request, response);
     }
 }
